requirements.txt
streamlit==1.44.1
google-generativeai==0.8.4
python-dotenv==1.0.1
pandas==2.1.4
ipython==8.19.0